requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.4.2
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.7.1
python-multipart==0.0.9